requirements.txt
Pillow==10.2.0
tkextrafont==0.6.3